+++ src/pages/topics/ui/index.ts
@@ -1,2 +1,2 @@
 export { StageTabs } from '../ui/stage-tabs';
-export { TopicsHeader } from '../ui/topics-header';
+export { TopicsHeader } from './topic-header';
